Catch exceptions in mysqli_close
If you try to close a connection that's already been in closed with mysql_close, it will return false. If you try to do it with mysqli_close, it will throw an exception (and return true, if you could get to that point). We try to mimic the behaviour of mysql_close by catching the exception.